requirements.txt
numpy
scipy
scikit-learn
pyyaml
pandas

# for tests
matplotlib